вроде сделал

+++ putstr.c
@@ -6,16 +6,16 @@
 /*   By: glavette <glavette@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/10 17:01:05 by glavette          #+#    #+#             */
-/*   Updated: 2021/11/10 17:12:48 by glavette         ###   ########.fr       */
+/*   Updated: 2021/11/11 19:28:39 by glavette         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "printf.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t			count;
-	const char		*string;
+	int			count;
+	const char	*string;
 
 	string = s;
 	count = 0;
@@ -29,8 +29,8 @@ size_t	ft_strlen(const char *s)
 
 int	ft_putstr(va_list ap)
 {
-	char*	a;
+	char	*a;
 
-	a = va_arg(ap, char*);
+	a = va_arg(ap, char *);
 	return (write (1, a, ft_strlen(a)));
 }
